+++ src/UI/UI_combobox.hpp
@@ -4,13 +4,13 @@
 #include "UI_x_boxes.hpp"
 
 #define UI_COMBOBOX_TYPE std::string("UIcombobox")
-typedef void (*UI_combobox_on_change) (std::string id, int selected);
+typedef void (*UI_combobox_on_change) (std::string id, size_t selected);
 
 class UI_combobox : public UI_x_boxes
 {
     public:
         UI_combobox();
-        UI_combobox(std::string id, std::string text, Color fg_color, Color bg_color, std::vector<std::string> options, int selected, UI_combobox_on_change on_change = nullptr);
+        UI_combobox(std::string id, std::string text, Color fg_color, Color bg_color, std::vector<std::string> options, size_t selected, UI_combobox_on_change on_change = nullptr);
         UI_combobox(std::string id, std::string text, Color fg_color, Color bg_color, std::vector<std::string> options, UI_combobox_on_change on_change = nullptr);
         UI_combobox(const json& j);
